refactor(LICENSE): Update license text
The commit includes changes to the LICENSE file. The license text was updated to be more concise and to align with the GNU General Public License Version 3, released on June 29, 2007. The changes ensure that the license document clearly outlines the permissions and responsibilities associated with sharing and modifying the software.

Signed-off-by: Schubert Anselme <schubert@anselm.es>