feat: Add initial files for hack directory
This commit adds the initial set of files for the `hack` directory. It includes a `.gitignore` file, a `.gitkeep` file for the `capi` subdirectory, and a `docker-compose.yaml` file for the `docker` subdirectory. The `.gitignore` file specifies which files should be ignored by Git, while the `docker-compose.yaml` file defines a Docker Compose configuration for a `k0s` container.

Signed-off-by: Schubert Anselme <schubert@anselm.es>